pyproject.toml
[project]
name = "claude-personal-assistant"
version = "0.1.0"
description = "A personal assistant agent that checks Notion ToDo lists, Dex CRM contacts, and sends reminder emails"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk>=0.1.19",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.3",
    "python-dotenv>=1.2.1",
]

[project.scripts]
assistant = "main:main"
